+++ package/src/__tests__/useGlobalGuards.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount, ReactWrapper } from 'enzyme';
-import { GuardFunction } from '../../types';
-import { GuardContext } from '../../contexts';
+import { GuardFunction } from '../types';
+import { GuardContext } from '../contexts';
 import { useGlobalGuards } from '../useGlobalGuards';
 
 const guardOne: GuardFunction = (to, from, next) => next();
